Bug 730142 - prepare TestRecordsChannel for symmetric TestFillingRecordsChannel.
